Private / Public API for RPM-Client module RemotingErrorHandlers (set) can be defined
