+++ integration-tests/src/test/java/com/canoo/dolphin/integration/ConnectionTest.java
@@ -1,6 +1,6 @@
 package com.canoo.dolphin.integration;
 
-import com.canoo.dolphin.client.ClientContext;
+import com.canoo.platform.client.ClientContext;
 import org.testng.Assert;
 import org.testng.annotations.Test;
