+++ integration-tests/src/test/java/com/canoo/dolphin/integration/PropertyControllerTest.java
@@ -1,7 +1,7 @@
 package com.canoo.dolphin.integration;
 
-import com.canoo.dolphin.client.ClientContext;
-import com.canoo.dolphin.client.ControllerProxy;
+import com.canoo.platform.client.ClientContext;
+import com.canoo.platform.client.ControllerProxy;
 import com.canoo.dolphin.integration.property.PropertyTestBean;
 import org.testng.Assert;
 import org.testng.annotations.Test;
