+++ platform-examples/process-monitor-sample/process-monitor-client-javafx/src/main/java/com/canoo/dolphin/samples/processmonitor/ProcessMonitorClient.java
@@ -15,7 +15,7 @@
  */
 package com.canoo.dolphin.samples.processmonitor;
 
-import com.canoo.dolphin.client.ClientContext;
+import com.canoo.platform.client.ClientContext;
 import com.canoo.dolphin.client.javafx.SimpleDolphinPlatformApplication;
 import javafx.application.Application;
 import javafx.application.Platform;
