+++ platform-examples/web-deployment-example/web-deployment-client/src/main/java/com/canoo/dolphin/webdeployment/client/view/MyView.java
@@ -1,6 +1,6 @@
 package com.canoo.dolphin.webdeployment.client.view;
 
-import com.canoo.dolphin.client.ClientContext;
+import com.canoo.platform.client.ClientContext;
 import com.canoo.dolphin.client.javafx.binding.FXBinder;
 
 import com.canoo.dolphin.client.javafx.view.AbstractFXMLViewBinder;
