+++ platform-extras/dolphin-platform-spring-controller-tester/src/main/java/com/canoo/dolphin/test/ControllerUnderTest.java
@@ -15,7 +15,7 @@
  */
 package com.canoo.dolphin.test;
 
-import com.canoo.dolphin.client.Param;
+import com.canoo.platform.client.Param;
 import com.canoo.platform.server.DolphinAction;
 import com.canoo.platform.server.DolphinController;
