+++ platform/dolphin-platform-rpm-client-javafx/src/main/java/com/canoo/dolphin/client/javafx/JavaFXConfiguration.java
@@ -15,7 +15,7 @@
  */
 package com.canoo.dolphin.client.javafx;
 
-import com.canoo.dolphin.client.ClientConfiguration;
+import com.canoo.platform.client.ClientConfiguration;
 import javafx.application.Platform;
 
 import java.net.URL;
