+++ platform/dolphin-platform-rpm-client-javafx/src/main/java/com/canoo/dolphin/client/javafx/SimpleDolphinPlatformApplication.java
@@ -1,6 +1,6 @@
 package com.canoo.dolphin.client.javafx;
 
-import com.canoo.dolphin.client.ClientInitializationException;
+import com.canoo.platform.client.ClientInitializationException;
 import com.canoo.platform.core.DolphinRuntimeException;
 import javafx.scene.control.Alert;
 import javafx.scene.control.ButtonType;
