+++ platform/dolphin-platform-rpm-client/src/main/java/com/canoo/dp/impl/client/ClientBeanBuilderImpl.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.canoo.dolphin.client.impl;
+package com.canoo.dp.impl.client;
 
 import com.canoo.dolphin.collections.ListChangeEvent;
 import com.canoo.dolphin.collections.ObservableList;
