+++ platform/dolphin-platform-rpm-client/src/main/java/com/canoo/dp/impl/client/ControllerProxyImpl.java
@@ -13,12 +13,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.canoo.dolphin.client.impl;
+package com.canoo.dp.impl.client;
 
-import com.canoo.dolphin.client.ControllerActionException;
-import com.canoo.dolphin.client.ControllerInitalizationException;
-import com.canoo.dolphin.client.ControllerProxy;
-import com.canoo.dolphin.client.Param;
+import com.canoo.platform.client.ControllerActionException;
+import com.canoo.platform.client.ControllerInitalizationException;
+import com.canoo.platform.client.ControllerProxy;
+import com.canoo.platform.client.Param;
 import com.canoo.dolphin.impl.InternalAttributesBean;
 import com.canoo.dolphin.impl.commands.CallActionCommand;
 import com.canoo.dolphin.impl.commands.DestroyControllerCommand;
