+++ platform/dolphin-platform-rpm-client/src/main/java/com/canoo/dp/impl/client/DolphinCommandHandler.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.canoo.dolphin.client.impl;
+package com.canoo.dp.impl.client;
 
 import com.canoo.impl.platform.core.Assert;
 import org.opendolphin.core.client.comm.AbstractClientConnector;
