+++ platform/dolphin-platform-rpm-client/src/main/java/com/canoo/platform/client/Param.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.canoo.dolphin.client;
+package com.canoo.platform.client;
 
 import com.canoo.impl.platform.core.Assert;
