+++ platform/dolphin-platform-rpm-client/src/test/java/com/canoo/dolphin/client/impl/TestDolphinCommandHandler.java
@@ -16,6 +16,7 @@
 package com.canoo.dolphin.client.impl;
 
 import com.canoo.dolphin.client.util.AbstractDolphinBasedTest;
+import com.canoo.dp.impl.client.DolphinCommandHandler;
 import org.opendolphin.core.client.ClientAttribute;
 import org.opendolphin.core.client.ClientDolphin;
 import org.opendolphin.core.comm.Command;
